+++ HISTORY.rst
@@ -9,6 +9,7 @@ PENDING
 * Brought back ``HISTORY.rst``.
 * Add ``mysqlparse.parse`` function.
 * Add ``six`` as a dependency to make writing Py2 and Py3 compatible code easier.
+* Add support for ``ALTER TABLE .. MODIFY [COLUMN]`` statements.
 
 0.1.3 (2016-02-20)
 ------------------